toolkit/mozapps/update/tests/chrome/test_0084_error_patchApplyFailure_verify_failed.xul

<?xml version="1.0"?>
<!--
/* Any copyright is dedicated to the Public Domain.
 * http://creativecommons.org/publicdomain/zero/1.0/
 */
-->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>

<window title="Update Wizard pages: error patching, download, and errors (partial failed and download complete verification failure)"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="runTestDefault();">
<script type="application/javascript"
        src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
<script type="application/javascript"
        src="utils.js"/>

<script type="application/javascript">
<![CDATA[

const TESTS = [ {
  pageid: PAGEID_ERROR_PATCHING,
  buttonClick: "next"
}, {
  pageid: PAGEID_DOWNLOADING,
  extraStartFunction: createContinueFile
}, {
  pageid: PAGEID_ERRORS,
  buttonClick: "finish",
  extraStartFunction: removeContinueFile
} ];

function runTest() {
  debugDump("entering");

  removeContinueFile();

  // Specify the url to update.sjs with a slowDownloadMar param so the ui can
  // load before the download completes.
  let slowDownloadURL = URL_HTTP_UPDATE_XML + "?slowDownloadMar=1";
  let patches = getLocalPatchString("partial", null, null, null, null, null,
                                    STATE_PENDING) +
                getLocalPatchString("complete", slowDownloadURL, "MD5",
                                    null, "1234",
                                    "false");
  let updates = getLocalUpdateString(patches, null, null, null,
                                     Services.appinfo.version,
                                     Services.appinfo.platformVersion, null,
                                     null, null, null, null, null, "false");

  writeUpdatesToXMLFile(getLocalUpdatesXMLString(updates), true);

  writeUpdatesToXMLFile(getLocalUpdatesXMLString(""), false);
  writeStatusFile(STATE_FAILED + ": " + CRC_ERROR);
  reloadUpdateManagerData();

  testPostUpdateProcessing();
}

]]>
</script>

<body xmlns="http://www.w3.org/1999/xhtml">
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test"></pre>
</body>
</window>
